Updated models, settings, forms & admin files
There was a lot done here. I've updated the models file to use a custom
User model (by the name of SMSUser) instead of the default one that
comes with django, I've also put in two other models by the name of
SMSPrice and Type.
I've also updated the settings file in-order to make it use the custom
user model instead of the default one, it also uses the external mysql
db instead of the default one, I've also install the commons app and the
rest_framework_swagger app. Forgot to add the rest_framework app, it
will quickly be done.
I've also created and updated an app level forms file that will be used
instead of the default django forms app.
I've also updated the admin file with a new customized class called
SMSUserAdmin that will be the one to use in the admin panel.
Hojjiin Itti fufa.