Done a lot
I've done a lot. I've defined a new apiviews.py file that is responsible
for generating the api views, defined a serializers.py file that is
responsible for converting the model representation to machine readable
format, aka json, re-configured my SMSUser and SMSPrice models to add a
verbose_name_plural and changed the foreign key relationships from null
to default, defined a new urls.py file in app level that will be
responsible for the urls of the commons app. I've changed the DATABASES
attribute in the settings.py file to include an OPTIONS attribute that
has set the foreign_key_checks to 0. I might have forgotten a few but
you can find them in the code.

Hojjiin Itti fufaa.