A lot done here
I've updated the settings.py file to include a rest_framework
authentication scheme, added a lot to the apiviews, including but not limited to, a way to update the SMSUser, (more can be found in the comments), added a few new attributes to the SMSUserSerialzer class, added a few urls to urls.py file and also changed a few others as well (more can be found in the comments).

I'm done with user creation, update, and login. Hojjiin itti fufaa.