successfully migrated the smsmessages table
The last commit (before this one) was done in order to have a snapshot
of the states the code was in...turns out I didn't need it b/c I didn't
need to delete the entire migrations folder, just remove the code that
used the model before it was migrated.
As for the rest: added the smsmessages model to the models.py file, migrate the smsmessages model (turns out you can give a custom name to the migration), added a homepage.html just so it won't return a 500 (with all the necessary css and images files), modified the urls.py to direct root requests to the fore-mentioned homepage.html, modified the views.py file to render the homepage.html file.