Made a few modifications to the SMSMessages model
as stated above, a few modifications have been made to the SMSMessages
model, they have been migrated, and the SMSMesages model has been added
to the admin panel. For some reason I've modified the forms.py file and
I expect it to bite my ass in the near future. The rest are simple pep8
modifications.