I did a lot on this one
I've done a lot on this one, namely:
1. updated settings.py file to include the notification app
2. updated the api/urls.py file to include the notification app's url
3. amended the url pattern in urls.py for each app
4. created an apiviews.py file for the notification app
5. in the notification/apiviews.py file I've overridden APIView to
customize the POST method
6. made several modifications to the models (details in the migration
files)
7. created a notification/sender.py file that will be responsible for
sending the sms to ethio-telecom's server (don't know if it should be in
the commons app b/c it will literally be common to all the apps)
8. added the rope module (used for, as far as I know, for updating
python references)
9. forgot all the rest
10. more detail can be found in the comments I've written and by
comparing the previous file versions.

Hojjin itti fufaa.