Updated several things
I've updated several things:
1. the api/urls.py file now contains a homepage that will be temporary
2. the commons/apiviews.py file import status from rest_framework, it
uses that to siginify statuses in its responses, it allows user update
on every part of the user attributes, with the exception of
sms_price.Should sms_price is attempted to be updated it will respond
with a 403 response code.
3. the notification/sender.py file is now modified to use the retry
function. The retry function will call the sender function.
4. notification/urls.py now has a url pattern of sending sms by the name
of send_sms.