finished the GET of SMSendView
I've finished the GET of SMSendView, it now returns the entire sent sms's for the requesting user or returns 404 (pagination might be necessary in the near future). I've also modified the notification/sender.py file with a comment on how to use the time module. I've added the url for viewsms that will call the get() method of the SMSendview class and finally added a space (somewhere) to the commons/models.py file.

Hojiin itti fufa.