Retry and a reasonable Response object complete
the update method defined in the SMSMessagesSerializer class needs an instance to run with, so that's been changed in the apiview.py file on sms_messages_serializer.update. The data attribute has also been removed.
The SMSSerializer now checks if it's valid before saving to db. And I've
decided that sender.py shall log every interaction with sms server and
that will be used as the necessary data.