working on login from ui
I've inserted some of the necessary ui html files that are going to be
used. They use bootstrap as a library (it was supposed to make it much
easier). currently it's not even logging in, that will have to be fixed.
All the rest has been done according to sendtex and the official django tutorial.

Hojjiin Itti Fufaa