Finished the ui portion of register.html
I've finished the ui portion of register.html. It currently sucks, is
not centered and has problems with smaller screens......but it works and
that's what I want at the moment.
I've also changed up the body portion of the whole site (aka in
base.html), it now has a new color.
I've changed the id of the actual nav item containing div. It's now navbar-main-navigation.