Forgot I changed the id of the input variables
Since I've changed the ID of the input variables to regusername,
regpassword and in the login.html file to logusername and logpassword,
the backend must reflect that change and now it has done so on the
login_request() function.
I've added jquery version 3.4.1 to the site as I will (probably) be
using that.
I've move the col-sm-12 definition to the div that contains the entire
row b/c it looks better on smaller devices over there.