Registration finally works
Finally registration works. It's not the most shining of all but it
works and thats what currently matters. It first checks if the username
exists within the database and if it already exists it will send a
rejection. But if it didn't it will successfully register the new user.
Now it has to login that same user and proceed on.

Hojjin Itti Fufaa