Dashboard is now customized to the user
the dashboard() function of the view.py file, now takes in the username as a parameter and
the dashboard.html template is not customized for the username.
messaging.html no longer has double quotes surrounding it.
the navbaritems.html now knows if the user has logged and what to
display in that case and if the user hasn't logged in/ is just viewing
the site.