Added a template for all the 404 responses
I've added a all404.html the 404 responses that will be shown to the
user.
I've also added a check_username_existence function to the
checking_utility.py module, it will responsible for checking if a
username exists after somebody has loaded the dashboard.
I've edited the ui/views module to reflect the above and now it works.

hojiin itti fufaa