I believe half way there to ajax update of dashboard
Fuck javascript

added a {% block scripts %} tag to base.html so as to simplify all the
rest of the calls to a javascript file that uses jquery.
added a {% block scripts %} tag to the dashboard.html file as well, and
inserted the script tag in there.
added a new url to ui:urls module. the one with the ajax/ is used to
by the ajax request to call the function defined.
in the ajax_dashboard_update function in ui:views module, it now return
a JsonResponse (imported that too) containing the total number of texts
sent (I still have to build a summation of this in the checking_utility
module) for the ajax request.

And FUCK ELPA.

Hojjin Itti Fufaa