A lot done here
modifications on the admin page to display a whole more detail on the
SMSMessages table than before, serveral makemigrations and migrates on
the models to include a new table name invoice and few modifications on
the previous ones, a new .py file for generating invoices, rename the
previous checking_utility.py to the better name of ui_utility.py, added
the invoice view to the views.py and url.py files.
Forgot the rest, perhaps there are comments in the code describing them,
perhaps not. Either way I'm way to late for deployment so....

Hojiin Itti fufaa.