Got diverted.
Whilst working on the invoice generation, got diverted to an issue with
dashboard that arises when a user hasn't sent any text messages. It now
checks if the return status is a 404 and responds to the ajax request
accordingly.

For the rest, I've forgot b/c I'm very nan marqina.