Improved the navbaritems file
I've improved the navbaritems file, navbaritems now uses the request.session.is_logged_in attribute to check if a user is logged in or not.
Improvements have been made on ui/invoice_generator.py to include the
the file upload location.
Improvements on ui/views.py file, specifically the invoice_generator
function. It now returns a render that passes a context that has a
username attribute. More to follow in the future.