Automatic invoice generation is now complete
I've made several changes to accomplish this:
1. a couple of migrations on the Invoice model.
2. modification on the ui/invoice_generator.py to add the month and year
it was created to the file name.
3. added the our company name to the invoice template that will be
rendered to pdf.
4. modified ui/invoices.html to symmetrically display all the invoices
created for a user.
5. modified ui.urls to reflect the changes on the ui.views.
I've also began the work on the settings view. It's currently on the
initial stages, meaning all it display is an under-construction image.
But it will be complete ASAP