A fucking lot has been done
It took way too much time, nearly a month, but I finally did it.
On the registration page it now shows an error message by using the
toast feature of bootstrap if the passwords and/or emails don't match.
In order to achieve that several modifications had to be made to the
register.html file and all the rest listed below. I'm surely gonna
regret this in the future but I simply don't have the moral or humna to
write a description about this so....future Gadd, BAAY'EE DHIIFAMAA