Collected all mentions of localhost into one var
it's named base_url and it contains the url that will be accessed. So
next time we can just change base_url for production.

P.S. for the future, perhaps putting base_url in the api/settings.py
file or even in the env file will do a good amount of work for us.