Modified the apiviews of notification app
I've modified the apiviews of the notification app, specifically on the
post method of the SMSView class. Now on the SMSMessagesSerializer's
data json object the sending user is retrieved from the auth token sent.
Note that you'll have to remind the users to send the json as object
with attributes as sms_number_to and sms_content.
I've also modified the sender module of the notification app,
specifically on the sender methods exception writing. It now adds a new
line after it writes.

Nearly forgot that I've figured out that the sms_messages_serializer has
an errors object that contains every detail on why it failed to be
valid, thus that will be written down on to a text file in server
whenever it occurs.

Please make sure it works on server before you give it to them.